Install script corrected to correctly perform reinstallation and print more information. #96884

+++ uninstall.sh
@@ -1,9 +1,19 @@
 #!/bin/sh
 
-# disable service
-systemctl stop pyssdp_server
-systemctl disable pyssdp_server
+if [ -e /etc/systemd/system/pyssdp_server.service ]
+then
+    echo "Stopping and disabling PySSDP Server service..."
+    # disable service
+    systemctl stop pyssdp_server
+    systemctl disable pyssdp_server
 
-# remove service source files from the system
-rm -rf /usr/lib/pyssdp_server
-rm -rf /etc/systemd/system/pyssdp_server.service
+
+    echo "Removing its files..."
+    # remove service source files from the system
+    rm -rf /usr/lib/pyssdp_server
+    rm -rf /etc/systemd/system/pyssdp_server.service
+
+    echo "PySSDP Server is now uninstalled from the system."
+else
+    echo "PySSDP Server is not installed."
+fi
